requirements.txt
numpy
h5py
requests
tqdm
pandas
tabulate
qdrant-client
pymilvus
weaviate-client
opensearch-py
elasticsearch
colorama
